Convert controller tests to requests specs
